Fixed canvas coordinate space to occupy entire element.
Involved changing board controller to redraw entire board to prevent
aliasing artifacts.

+++ src/main/java/demo/client/local/game/tools/Size.java
@@ -96,7 +96,7 @@ public class Size {
         size.blockSize = size.coordWidth / 4;
       }
       else {
-        size.blockSize = size.coordHeight / HEIGHT;
+        size.blockSize = ((double) size.coordHeight) / HEIGHT;
       }
     }
   }
